+++ serviceImpl/wei/src/main/java/cn/qihangerp/module/open/wei/domain/WeiOrder.java
@@ -41,12 +41,12 @@ public class WeiOrder implements Serializable {
     /**
      * 秒级时间戳
      */
-    private Integer createTime;
+    private Long createTime;
 
     /**
      * 秒级时间戳
      */
-    private Integer updateTime;
+    private Long updateTime;
 
     /**
      *
